When checking git version pass the --git-dir in to fix out of source tree build
When the rebuild gitversion script is run, it does not
get the correct location if it uses a CMAKE source var because it
is run outside of cmake config (during make). Generate a param
from the true source dit during config and pass that to the script
as an arg.