assert_same<T1,T2>() generates an error with T1,T2 in the error msg
This is a utility that can be useful for diagnosing type related problems.
When std::is_same<T1,T2> is used, an assertion can be thrown to halt
compilation, but the user can't see the types being compared.
This utility is templated over the types so that the error message
is decorated with the types tested and helps the user see what's wrong.