Fix execution of tests, #ifdef for POSIX/STDTHREAD and code style
Many tweaks to ensure that when CMake changes are made to
POSIX/STDTHREAD option, the correct #defines are used and the correct
tests are run.

Add STDTHREAD option to gtest macro and allow a test to be specified
for one or more thread libraries at the same time.

Code formatting fixes.